requirements.txt
# === Core ===
streamlit>=1.30.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# === Data ===
duckdb>=0.9.0
pandas>=2.1.0
pyarrow>=14.0.0

# === Geo / Map ===
geopandas>=0.14.0
folium>=0.15.0
streamlit-folium>=0.17.0
osmnx>=1.9.0
networkx>=3.2.0
shapely>=2.0.0
branca>=0.7.0

# === Visualization ===
plotly>=5.18.0
matplotlib>=3.8.0

# === LLM / RAG ===
google-generativeai>=0.4.0
langchain>=0.1.0
langchain-google-genai>=0.0.6
faiss-cpu>=1.7.4
sentence-transformers>=2.3.0

# === HTTP / API ===
httpx>=0.26.0
requests>=2.31.0

# === Utilities ===
pydantic>=2.5.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
